refactor: move query_finalized_custodian_cells to rpc_client
