+++ crates/block-producer/src/rpc_client.rs
@@ -5,21 +5,26 @@ use crate::types::CellInfo;
 use anyhow::{anyhow, Result};
 use async_jsonrpc_client::{HttpClient, Output, Params as ClientParams, Transport};
 use ckb_types::prelude::Entity;
-use gw_common::H256;
+use gw_common::{CKB_SUDT_SCRIPT_ARGS, H256};
 use gw_generator::RollupContext;
 use gw_jsonrpc_types::ckb_jsonrpc_types::{self, BlockNumber, Uint32};
 use gw_types::{
     bytes::Bytes,
     core::ScriptHashType,
     packed::{
-        Block, CellOutput, DepositionLockArgs, DepositionLockArgsReader, DepositionRequest,
-        NumberHash, OutPoint, Script, StakeLockArgs, StakeLockArgsReader, Transaction,
+        Block, CellOutput, CustodianLockArgs, CustodianLockArgsReader, DepositionLockArgs,
+        DepositionLockArgsReader, DepositionRequest, NumberHash, OutPoint, Script, StakeLockArgs,
+        StakeLockArgsReader, Transaction,
     },
     prelude::*,
 };
 use serde::de::DeserializeOwned;
 use serde_json::{from_value, json};
-use std::{collections::HashSet, time::Duration};
+
+use std::{
+    collections::{HashMap, HashSet},
+    time::Duration,
+};
 
 pub const DEFAULT_QUERY_LIMIT: usize = 1000;
 
@@ -87,6 +92,40 @@ fn parse_deposit_request(
         .sudt_script_hash(sudt_script_hash.pack())
         .build();
     Some(request)
+}
+
+#[derive(Debug)]
+pub struct WithdrawalsAmount {
+    pub capacity: u64,
+    pub sudt: HashMap<[u8; 32], u128>,
+}
+
+impl Default for WithdrawalsAmount {
+    fn default() -> Self {
+        WithdrawalsAmount {
+            capacity: 0,
+            sudt: Default::default(),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct CollectedCustodianCells {
+    pub cells_info: Vec<CellInfo>,
+    pub capacity: u64,
+    pub sudt: HashMap<[u8; 32], u128>,
+    pub fullfilled_sudt_script: HashMap<[u8; 32], Script>,
+}
+
+impl Default for CollectedCustodianCells {
+    fn default() -> Self {
+        CollectedCustodianCells {
+            cells_info: Default::default(),
+            capacity: 0,
+            sudt: Default::default(),
+            fullfilled_sudt_script: Default::default(),
+        }
+    }
 }
 
 #[derive(Clone)]
@@ -622,6 +661,143 @@ impl RPCClient {
         };
 
         Ok(unlocked_cell.map(unlocked_cell_info))
+    }
+
+    pub async fn query_finalized_custodian_cells(
+        &self,
+        withdrawals_amount: &WithdrawalsAmount,
+        last_finalized_block_number: u64,
+    ) -> Result<CollectedCustodianCells> {
+        let rollup_context = &self.rollup_context;
+
+        let parse_sudt_amount = |cell: &Cell| -> Result<u128> {
+            if cell.output.type_.is_none() {
+                return Err(anyhow!("no a sudt cell"));
+            }
+
+            gw_types::packed::Uint128::from_slice(&cell.output_data.as_bytes())
+                .map(|a| a.unpack())
+                .map_err(|e| anyhow!("invalid sudt amount {}", e))
+        };
+
+        let custodian_lock = Script::new_builder()
+            .code_hash(rollup_context.rollup_config.custodian_script_type_hash())
+            .hash_type(ScriptHashType::Type.into())
+            .args(rollup_context.rollup_script_hash.as_slice().pack())
+            .build();
+
+        let search_key = SearchKey {
+            script: ckb_types::packed::Script::new_unchecked(custodian_lock.as_bytes()).into(),
+            script_type: ScriptType::Lock,
+            filter: None,
+        };
+        let order = Order::Desc;
+        let limit = Uint32::from(DEFAULT_QUERY_LIMIT as u32);
+
+        let mut collected = CollectedCustodianCells::default();
+        let mut cursor = None;
+
+        while collected.capacity < withdrawals_amount.capacity
+            || collected.fullfilled_sudt_script.len() < withdrawals_amount.sudt.len()
+        {
+            let cells: Pagination<Cell> = to_result(
+                self.indexer_client
+                    .request(
+                        "get_cells",
+                        Some(ClientParams::Array(vec![
+                            json!(search_key),
+                            json!(order),
+                            json!(limit),
+                            json!(cursor),
+                        ])),
+                    )
+                    .await?,
+            )?;
+
+            if cells.last_cursor.is_empty() {
+                return Err(anyhow!("no finalized custodian cell"));
+            }
+            cursor = Some(cells.last_cursor);
+
+            for cell in cells.objects.into_iter() {
+                let args = cell.output.lock.args.clone().into_bytes();
+                let custodian_lock_args = match CustodianLockArgsReader::verify(&args[32..], false)
+                {
+                    Ok(()) => CustodianLockArgs::new_unchecked(args.slice(32..)),
+                    Err(_) => continue,
+                };
+
+                if custodian_lock_args.deposition_block_number().unpack()
+                    > last_finalized_block_number
+                {
+                    continue;
+                }
+
+                // Collect sudt
+                if let Some(json_script) = cell.output.type_.clone() {
+                    let sudt_type_script = {
+                        let script = ckb_types::packed::Script::from(json_script);
+                        Script::new_unchecked(script.as_bytes())
+                    };
+
+                    let sudt_type_hash = sudt_type_script.hash();
+                    if sudt_type_hash != CKB_SUDT_SCRIPT_ARGS {
+                        // Already collected enough sudt amount
+                        let fullfilled_sudt_script = &mut collected.fullfilled_sudt_script;
+                        if fullfilled_sudt_script.contains_key(&sudt_type_hash) {
+                            continue;
+                        }
+
+                        // Not targed withdrawal sudt
+                        let withdrawal_amount = match withdrawals_amount.sudt.get(&sudt_type_hash) {
+                            Some(amount) => amount,
+                            None => continue,
+                        };
+
+                        let sudt_amount = match parse_sudt_amount(&cell) {
+                            Ok(amount) => amount,
+                            Err(_) => {
+                                log::error!("invalid sudt amount, out_point: {:?}", cell.out_point);
+                                continue;
+                            }
+                        };
+
+                        let collected_amount = collected.sudt.entry(sudt_type_hash).or_insert(0);
+                        *collected_amount = collected_amount.saturating_add(sudt_amount);
+
+                        if *collected_amount >= *withdrawal_amount {
+                            fullfilled_sudt_script
+                                .insert(sudt_type_hash.to_owned(), sudt_type_script);
+                        }
+                    }
+                }
+
+                // Collect capacity
+                let out_point = {
+                    let out_point: ckb_types::packed::OutPoint = cell.out_point.into();
+                    OutPoint::new_unchecked(out_point.as_bytes())
+                };
+
+                let output = {
+                    let output: ckb_types::packed::CellOutput = cell.output.into();
+                    CellOutput::new_unchecked(output.as_bytes())
+                };
+
+                collected.capacity = collected
+                    .capacity
+                    .saturating_add(output.capacity().unpack());
+
+                let info = CellInfo {
+                    out_point,
+                    output,
+                    data: cell.output_data.into_bytes(),
+                };
+
+                collected.cells_info.push(info);
+            }
+        }
+
+        Ok(collected)
     }
 
     pub async fn get_transaction(&self, tx_hash: H256) -> Result<Option<Transaction>> {
